+++ examples/sycl/06_pvc_flash_attention/pvc_flash_attn_runner.hpp
@@ -417,11 +417,11 @@ template <class GemmKernel, bool isVarLen> struct ExampleRunner {
     stride_V = cutlass::make_cute_packed_stride(StrideV{}, cute::make_shape(head_size_vo, seq_len_kv, batch * num_heads_kv));
     stride_O = cutlass::make_cute_packed_stride(StrideO{}, cute::make_shape(seq_len_qo, head_size_vo, batch * num_heads_q));
 
-    block_Q.reset(batch * num_heads_q * seq_len_qo * head_size_qk);
-    block_K.reset(batch * num_heads_kv * seq_len_kv * head_size_qk);
-    block_V.reset(batch * num_heads_kv * seq_len_kv * head_size_vo);
-    block_O.reset(batch * num_heads_q * seq_len_qo * head_size_vo);
-    block_ref_O.reset(batch * num_heads_q * seq_len_qo * head_size_vo);
+    block_Q.reset(static_cast<std::size_t>(batch) * num_heads_q * seq_len_qo * head_size_qk);
+    block_K.reset(static_cast<std::size_t>(batch) * num_heads_kv * seq_len_kv * head_size_qk);
+    block_V.reset(static_cast<std::size_t>(batch) * num_heads_kv * seq_len_kv * head_size_vo);
+    block_O.reset(static_cast<std::size_t>(batch) * num_heads_q * seq_len_qo * head_size_vo);
+    block_ref_O.reset(static_cast<std::size_t>(batch) * num_heads_q * seq_len_qo * head_size_vo);
 
     initialize_block(block_Q, seed + 2023);
     initialize_block(block_K, seed + 2022);
